Rework to drastically improve runtime for large number of replacements
Doing the replacement page-by-page became slow with large number of
items.

We switch this so that we do the final combining of the Xml-structure
only once at the end.

This may require some more memory as we produce the full string in
memory and from this produce the resulting Xml object, however it
seems to work fine for hundreds of lines with 256m memory.